Adding Frankies files
Uploading frankies files to the repo. Current state of code is that it
should work but hasn’t been tested for one RPM. Hasn’t been implemented
to work with the RPM vector yet. Kind of sketchy use of T2m to find
initial gamma for use in m_dot calculation at the very beginning.